requirements.txt
pip
rich
ollama
python-dotenv
deep-translator
